% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRepsd.R
\docType{methods}
\name{CRepsd}
\alias{CRepsd}
\title{To combine the replicates of normalised and standardized intensities and to get
standard deviation}
\usage{
\S4method{CRepsd}{gnsdt}(gnsdt)
}
\arguments{
\item{gnsdt}{object of the S4 Class \code{gnsdt}}
}
\value{
object of Class \code{"gnsdt"} and save the results in slot cData
}
\description{
Function \code{CRepsd} works on the output of the function \code{readGNS}, which is
an object of Class \code{gnsdt} and populate them in an object of Class \code{gnsdt}
}
\details{
Allows the user to compute normalization and standardization of intensites, to
combine the replicates of normalized  and standardized intensities,to compute the standard
deviation (S.D.) of normalized and standardized intensities within replicates, to compute the
S.D. of intensities of control and test raw data intensity within replicates and save them in
different slots: (1)nsData: Normalized and standardized intensities (2)cData: to save
combined replicates of normalized and standardized intensities (3) sdnsData: to contain the
standard deviation of normalized and standardized intensities within replicates (4) sdData: to
contain the standard deviation of control and test line raw intensity data within replicates
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "Imager.csv", package = "ReadqPCR")
data <- readGNS(file = file1)
## to visualise all the data with all the slots
data
## to visualise the initial data
slot(data,"iData")
## to read the file
crep1 <- CRepsd(data)
## to visualize the overall results
crep1
## to visualize the normalized and standardized intensities
slot(crep1,"nsData")
## to visualize the combined replicates
slot(crep1,"cData")
## to visualize the standard deviation of combined replicates
slot(crep1,"sdnsData")
## to visualize the standard deviation of raw intensity data
slot(crep1,"sdData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{gnsdt}
