% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntensity.R
\name{getIntensity}
\alias{getIntensity}
\title{To get the final extracted intensity of control line and test line signals,
after baseline correction}
\usage{
getIntensity(imagefile, cut1 = 8, cut2 = 25, pi = 1.2, fp = TRUE)
}
\arguments{
\item{imagefile}{.jpeg image of the lateral flow assay}

\item{cut1}{first value on X-axis (index) to cut the control line or test line
peak, as obtained from the function readIGNScutprocess}

\item{cut2}{second value on x-axis (index) to cut the control line or test line
peak as obtained from the readIGNScutprocess}

\item{pi}{the value of intensity to be applied as baseline correction for control
line or test line peaks (signals). These should be the same values, as applied in
the readIGNScutprocess}

\item{fp}{TRUE or FALSE. When TRUE, one can implement for first peak control line.
When FALSE; one can implement for second peak test line.}
}
\value{
baseline corrected intensities of control line and test line
}
\description{
Function \code{getIntensity} reads the .jpeg files of the lateral
flow assay and returns the extracted intensity of the control line and test
line signals
}
\details{
Allows the user to get the baseline corrected values of intensities of
control line or test line
}
\examples{
#'## to read and see complete plot of signal
file3 <- system.file("exData", "Serum30nM_2.JPG", package = "GNSplex")
readIGNSplot(file3)
## to read and cut the first peak (control line signal) and to do baseline
##correction
file3 <- system.file("exData", "Serum30nM_2.JPG", package = "GNSplex")
readIGNScutprocess(imagefile = file3, fp = TRUE, cut1 = 3, cut2 = 13, pi = 2.0)
## To cut the second peak (test line signal) and to do baseline correction
readIGNScutprocess(imagefile = file3, fp = FALSE, cut1 = 17, cut2 = 30, pi = 2.0)
## To get the baseline corrected intensity of control line (peak 1)
getIntensity(imagefile = file3, cut1 = 3, cut2 = 13, pi = 2.0)
## to get the baseline corrected intensity of test line (peak 2)
getIntensity(imagefile = file3, cut1 = 17, cut2 = 30, pi = 2.0, fp = FALSE)
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{gnsdt}
