% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnsConc.R
\name{gnsConc}
\alias{gnsConc}
\title{To compute the concentration}
\usage{
gnsConc(y, m, c)

gnsConc(y, m, c)
}
\arguments{
\item{Y}{Normalized or standardized intensity}

\item{M}{slope}

\item{C}{intercept}

\item{Y}{Normalized or standardized intensity}

\item{M}{slope}

\item{C}{intercept}
}
\value{
X value, which is the concentration

X value, which is the concentration
}
\description{
Function \code{gnsConc} computes the concentration on the basis of
formula Y = MX + C, from nromalized or standardized intensity

Function \code{gnsConc} computes the concentration on the basis of
formula Y = MX + C, from nromalized or standardized intensity
}
\details{
Allows the user to compute concentration from intensity

Allows the user to compute concentration from intensity
}
\examples{
## m and c are the values obtained from the linear model equation
## y is the normalized or standardized intensity for which
##concentration is to be calculated
concentration <- gnsConc(0.45,0.0213,0.346)
concentration
## m and c are the values obtained from the linear model equation
## y is the normalized or standardized intensity for which
##concentration is to be calculated
concentration <- gnsConc(0.45,0.0213,0.346)
concentration
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}

Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{gnsdt}
