% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnsConf.R
\docType{methods}
\name{gnsConf}
\alias{gnsConf}
\title{To compute the confidence interval of combined replicates of normalized and standardized
intensities}
\usage{
\S4method{gnsConf}{gnsdt}(gnsdt)
}
\arguments{
\item{gnsdt}{an output of function CRepsd, which is an object of class \code{gnsdt}}
}
\value{
object of Class \code{"gnsdt"} and save the results in slot confData of an object
of class \code{gnsdt}
}
\description{
Function \code{gnsConf} works on the output of function CRepsd, which is an
object of S4 class \Code{gnsdt} and computes the confidence interval of combined replicates
of normalized and standardized intensities and save them in a new slot \code{confData} of
Class \code{gnsdt}
}
\details{
Allows the user to compute confidence interval of combined replicates of normalized and
standardized intensities and populate an object of class \code{gnsdt}
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{gnsdt}
