% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnscor.R
\docType{methods}
\name{gnscor}
\alias{gnscor}
\title{To compute the correlation of normalized and standardized intensities with
respect to their corresponding predicted values of linearly fitted model}
\usage{
\S4method{gnscor}{gnsdt}(gnsdt)
}
\arguments{
\item{gnsdt}{an output of function CRepsd, which is an object of the class
\Code{gnsdt}}
}
\value{
object of Class \code{"gnsdt"} and save the results of correlation in
slot corData
}
\description{
Function \code{gnscor} works on the output of function CRepsd, which is
an object of class \code{gnsdt} and computes the correlation between normalized
intensity with predicted values of normalized intensities from lm model as well as
correlation between standardized intensities with predicted values of standardized
intensities from lm model. It saves the final results of correlation in a new slot
corData of an object of class \Code{gnsdt}
}
\details{
Allows the user to compute the correlation and saves the results in the new
slot corData of an object of class \code{gnsdt}
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "Imager.csv", package = "ReadqPCR")
data <- readGNS(file = file1)
## To combine the replicates
crep1 <- CRepsd(data)
## To compute the correlation
cor1 <- gnscor(crep1)
## to visualize all the results
cor1
## to visualize the results of correlation only
slot(cor1,"corData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{gnsdt}
