% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{gnsdt-class}
\alias{gnsdt-class}
\title{S4 Class gnsdt to contain the data}
\description{
S4 Class gnsdt to contain the data
}
\section{Slots}{

\describe{
\item{\code{iData}}{contains the initial data, except concentration, in the form of an object
of Class \code{data.frame}}

\item{\code{nsData}}{contains the normalized and standardized intensity data in the form of an obkect
of Class \code{data.frame}}

\item{\code{cData}}{contains the result of the function \code{CombineReps} in the form of an object
of Class \code{data.frame}}

\item{\code{cData}}{contains the combined replicates in the form of an object of Class \code{data.frame}}

\item{\code{sdnsData}}{contains the standard deviation of normalized and standardized intensities
within repliactes in the for of an object of Class \code{data.frame}}

\item{\code{sdData}}{contains the stadard deviation of raw intensities within replicates in the form of an
object of Class \code{data.frame}}

\item{\code{confData}}{contains the confidence intervals in the form of an object of Class \code{data.frame}}

\item{\code{lodData}}{contains the result of LOB, LOD and LOQ in the form of an object of Class
\code{data.frame}}

\item{\code{corData}}{contains the result of correlation of normalized and standardized intensities
with respect to their predicted values by linear model in the form of an object of Class
\code{data.frame}}

\item{\code{conc}}{contains the data of concentration in the form of an object of class \code{data.frame}}
}}

