% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnsmodel.R
\docType{methods}
\name{gnsmodel}
\alias{gnsmodel}
\title{To generate the linear fitted plots of lm(Normalized intensity ~ Concentration)
and lm(Standardized intensity ~ Cocentration)}
\usage{
\S4method{gnsmodel}{gnsdt}(gnsdt, Normal = TRUE)
}
\arguments{
\item{gnsdt}{output of function \code{CRepsd}, which is an object of class \Code{gnsdt}}

\item{Normal}{choose between TRUE or FALSE, where TRUE is default. When TRUE, then returns
the plots of lm(normalized intensity~Concentration) and when FALSE, then returns the
plots of lm(standardized intensity~Concentration)}
}
\value{
linearly fitted plots
}
\description{
Function \code{gnsmodel} generates the plots of linear model of
normalized and standardized intensities vs Concentration. It works on the output
of CRepsd, which is an object of class \code{gnsdt} and utilized the combined
replicates of normalized and standardized intensities to fit linearly with
concentration data.
}
\details{
Allows the user to get the linearly fitted plots of normalized and standardized
intensities after combining replicates with respect to concentration
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "Imager.csv", package = "ReadqPCR")
data <- readGNS(file = file1)
## to combine the replicates of normalized and standardized intensities
crep1 <- CRepsd(data)
## to get the linearly fitted plot of normalized intensities vs concentration
gnsmodel(crep1)
## to get the linearly fitted plots of standardized intensities vs concentration
gnsmodel(crep1, Normal = FALSE)
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{gnsdt}
