% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNSplex.gui.R
\name{GNSplex.gui}
\alias{GNSplex.gui}
\title{Run the graphical user interface for analysis of the data of lateral
flow assay}
\usage{
GNSplex.gui()
}
\description{
Run the graphical user interface
}
\examples{
GNsplex.gui()
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
