% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGNS.R
\name{readGNS}
\alias{readGNS}
\title{To read the raw intensity data files}
\usage{
readGNS(file, type = ".csv")
}
\arguments{
\item{file}{raw data file of intensity to be read in}

\item{type}{choose between .csv or .txt, where .csv is default}
}
\value{
object of Class \code{"gnsdt"} and save the initial data in slot iData
}
\description{
Function \code{readGNS} reads the raw intensity files (.csv or .txt) and populate
them in an object of Class \code{gnsdt}
}
\details{
Allows the user to read in the raw intensity data and save them in S4 class
\code{gnsdt}
}
\examples{
##To read the .csv file
file1 <- system.file("exData", "Imager.csv", package = "ReadqPCR")
data <- readGNS(file = file1)
## to visualise all the data with all the slots
data
## to visualise the initial data
slot(data,"iData")
## to read the .txt file
file2 <- system.file("exData", "Iphone_R.txt", package = "ReadqPCR")
data2 <- readGNS(file = file2, type = ".txt")
## to visualize all the data with all the slots
data2
## to visualise the initial data
data2(data2,"iData")
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{gnsdt}
