% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIGNScutprocess.R
\name{readIGNScutprocess}
\alias{readIGNScutprocess}
\title{To read the .jpeg image of lateral flow assay and apply cuts to cut the
control line and test line peaks of the signal intensity vs index}
\usage{
readIGNScutprocess(imagefile, fp = TRUE, cut1 = 2, cut2 = 18,
  pi = 1.56)
}
\arguments{
\item{imagefile}{the .jpeg file of lateral flow assay}

\item{fp}{either TRUE or FALSE}

\item{cut1}{the first value on x-axis (index) to cut the peak}

\item{cut2}{the second value on x-axis (index) to cut the peak}

\item{pi}{the intensity (y-axis) to get the processed signal. It's a cutoff intensity
to do the baseline correction}
}
\value{
The plots of control line and test line peaks
}
\description{
Function \code{readIGNScutprocess} will provide the values, which can be
used to extract the intensities of control line and test line peaks
}
\details{
Allows the user to read in the .jpeg image of lateral flow assay and returns
the cutted and baseline corrected plots of signals of control line and test line.
}
\examples{
## to read and see complete plot of signal
file3 <- system.file("exData", "Serum30nM_2.JPG", package = "GNSplex")
readIGNSplot(file3)
## to read and cut the first peak (control line signal) and to do baseline
##correction
file3 <- system.file("exData", "Serum30nM_2.JPG", package = "GNSplex")
readIGNScutprocess(imagefile = file3, fp = TRUE, cut1 = 3, cut2 = 13, pi = 2.0)
## To cut the second peak (test line signal) and to do baseline correction
readIGNScutprocess(imagefile = file3, fp = FALSE, cut1 = 17, cut2 = 30, pi = 2.0)
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{gnsdt}
