% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIGNSplot.R
\name{readIGNSplot}
\alias{readIGNSplot}
\title{To read the image file}
\usage{
readIGNSplot(Imagefile)
}
\arguments{
\item{Imagefile}{well cutted .jpeg image of lateral flow strip}
}
\value{
plot of signal intensity vs index
}
\description{
Function \code{readIGNSplot} reads the raw image file and generates the
plot of signal (intensity vs index)
}
\details{
Allows the user to read in the .jpeg image of lateral flow strip and
generates the signal in the form of the plot (intensity vs index)
}
\examples{
##To read and plot the signal of lateral flow assay strip
file <- system.file("exData", "Serum30nM_2.JPG", package = "GNSplex")
readIGNSplot(file)
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{gnsdt}
